--- ha-batch/.gitattributes
@@ -1,3 +1,0 @@
-# /gitignore
-*.bat           text eol=crlf
-*.sh           text eol=lf
