+++ ha-batch/README.md
@@ -12,7 +12,7 @@
     - 概要：特定の月の健康情報を検索し、CSVに変換し、S3へのアップロードを行うバッチ
 
 ## ログ
-- 文字コード：UTF-8
+- 文字コード：MS932
 - 配置先：
     - local環境：C:/app/logs/
     - EC2環境：/var/log/app/
