+++ ha-batch/bat/monthlyHealthInfoSummary.bat
@@ -5,17 +5,22 @@ rem 月次健康情報集計バッチ
 rem 引数1:処理対象年月(YYYYMM)
 rem ----------------------------------------------------------------------------------------
 
+echo ------------------------------------------------------------------------
+echo START monthlyHealthInfoSummary.bat
+echo ------------------------------------------------------------------------
+
 rem 共通シェルの読込
 call common.bat
 
-cd %BASE_DIR%\ha-batch\target\classes
+cd %BASE_DIR%\ha-batch\target
 
-rem -m
-set DATE_OPTION=%1
 rem 処理対象年月YYYYMM
-set TARGET_DATE=%2
+set DATE_OPTION_VALUE=%1
 
-rem システム環境変数のJava(JAVA_HOME)だとeclipseでコンパイルしたときのJavaのversionと異なるので、直接、eclipseのJavaで実行する
-%JAVA% -cp %CLASSPATH% %MAIN% MonthlyHealthInfoSummaryBatch %DATE_OPTION% %TARGET_DATE%
+java -jar -Dspring.profiles.active=%ENV% -Dspring.batch.job.names=ｍonthlyHealthInfoSummaryBatchJob %JAR_FILE% m=%DATE_OPTION_VALUE%
+
+echo ------------------------------------------------------------------------
+echo END monthlyHealthInfoSummary.bat
+echo ------------------------------------------------------------------------
 
 cd %~dp0
