+++ ha-batch/sql/develop.sql
@@ -1,0 +1,86 @@
+-- SpringBatchのメタデータを作成するためのSQL
+DROP TABLE IF EXISTS batch_step_execution_seq;
+DROP TABLE IF EXISTS batch_step_execution_context;
+DROP TABLE IF EXISTS batch_step_execution;
+DROP TABLE IF EXISTS batch_job_seq;
+DROP TABLE IF EXISTS batch_job_execution_seq;
+DROP TABLE IF EXISTS batch_job_execution_params;
+DROP TABLE IF EXISTS batch_job_execution_context;
+DROP TABLE IF EXISTS batch_job_execution;
+DROP TABLE IF EXISTS batch_job_instance;
+
+CREATE TABLE IF NOT EXISTS BATCH_STEP_EXECUTION_SEQ (ID BIGINT NOT NULL);
+INSERT INTO BATCH_STEP_EXECUTION_SEQ values(0);
+CREATE TABLE IF NOT EXISTS BATCH_JOB_EXECUTION_SEQ (ID BIGINT NOT NULL);
+INSERT INTO BATCH_JOB_EXECUTION_SEQ values(0);
+CREATE TABLE IF NOT EXISTS BATCH_JOB_SEQ (ID BIGINT NOT NULL);
+INSERT INTO BATCH_JOB_SEQ values(0);
+
+CREATE TABLE IF NOT EXISTS BATCH_JOB_INSTANCE  (
+  JOB_INSTANCE_ID BIGINT  PRIMARY KEY ,
+  VERSION BIGINT,
+  JOB_NAME VARCHAR(100) NOT NULL ,
+  JOB_KEY VARCHAR(2500)
+);
+
+CREATE TABLE IF NOT EXISTS BATCH_JOB_EXECUTION_PARAMS  (
+  JOB_EXECUTION_ID BIGINT NOT NULL ,
+  TYPE_CD VARCHAR(6) NOT NULL ,
+  KEY_NAME VARCHAR(100) NOT NULL ,
+  STRING_VAL VARCHAR(250) ,
+  DATE_VAL DATETIME DEFAULT NULL ,
+  LONG_VAL BIGINT ,
+  DOUBLE_VAL DOUBLE PRECISION ,
+  IDENTIFYING CHAR(1) NOT NULL
+);
+
+CREATE TABLE IF NOT EXISTS BATCH_JOB_EXECUTION  (
+  JOB_EXECUTION_ID BIGINT  PRIMARY KEY ,
+  VERSION BIGINT,
+  JOB_INSTANCE_ID BIGINT NOT NULL,
+  CREATE_TIME TIMESTAMP NOT NULL,
+  START_TIME TIMESTAMP DEFAULT NULL,
+  END_TIME TIMESTAMP DEFAULT NULL,
+  STATUS VARCHAR(10),
+  EXIT_CODE VARCHAR(20),
+  EXIT_MESSAGE VARCHAR(2500),
+  LAST_UPDATED TIMESTAMP,
+  JOB_CONFIGURATION_LOCATION VARCHAR(2500) NULL,
+  constraint JOB_INSTANCE_EXECUTION_FK foreign key (JOB_INSTANCE_ID)
+  references BATCH_JOB_INSTANCE(JOB_INSTANCE_ID)
+) ;
+
+CREATE  TABLE IF NOT EXISTS BATCH_STEP_EXECUTION  (
+  STEP_EXECUTION_ID BIGINT  PRIMARY KEY ,
+  VERSION BIGINT NOT NULL,
+  STEP_NAME VARCHAR(100) NOT NULL,
+  JOB_EXECUTION_ID BIGINT NOT NULL,
+  START_TIME TIMESTAMP NOT NULL ,
+  END_TIME TIMESTAMP DEFAULT NULL,
+  STATUS VARCHAR(10),
+  COMMIT_COUNT BIGINT ,
+  READ_COUNT BIGINT ,
+  FILTER_COUNT BIGINT ,
+  WRITE_COUNT BIGINT ,
+  READ_SKIP_COUNT BIGINT ,
+  WRITE_SKIP_COUNT BIGINT ,
+  PROCESS_SKIP_COUNT BIGINT ,
+  ROLLBACK_COUNT BIGINT ,
+  EXIT_CODE VARCHAR(20) ,
+  EXIT_MESSAGE VARCHAR(2500) ,
+  LAST_UPDATED TIMESTAMP,
+  constraint JOB_EXECUTION_STEP_FK foreign key (JOB_EXECUTION_ID)
+  references BATCH_JOB_EXECUTION(JOB_EXECUTION_ID)
+) ;
+
+CREATE TABLE IF NOT EXISTS BATCH_JOB_EXECUTION_CONTEXT  (
+  JOB_EXECUTION_ID BIGINT PRIMARY KEY,
+  SHORT_CONTEXT VARCHAR(2500) NOT NULL,
+  SERIALIZED_CONTEXT BLOB
+) ;
+
+CREATE TABLE IF NOT EXISTS BATCH_STEP_EXECUTION_CONTEXT  (
+  STEP_EXECUTION_ID BIGINT PRIMARY KEY,
+  SHORT_CONTEXT VARCHAR(2500) NOT NULL,
+  SERIALIZED_CONTEXT BLOB
+) ;
