+++ ha-batch/src/main/resources/application-local.yml
@@ -1,0 +1,7 @@
+spring:
+  # DB設定(ha-dbの設定が取れないためここで定義)
+  datasource: 
+    url: jdbc:mysql://localhost:3306/work3g
+    username: root
+    password: admin
+    driver-class-name: com.mysql.cj.jdbc.Driver
