--- ha-batch/src/main/resources/batch-context.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans
-  xmlns="http://www.springframework.org/schema/beans"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xmlns:context="http://www.springframework.org/schema/context"
-  xmlns:mvc="http://www.springframework.org/schema/mvc"
-  xsi:schemaLocation="http://www.springframework.org/schema/mvc
-                      http://www.springframework.org/schema/mvc/spring-mvc.xsd
-                      http://www.springframework.org/schema/beans
-                      http://www.springframework.org/schema/beans/spring-beans.xsd
-                      http://www.springframework.org/schema/context
-                      http://www.springframework.org/schema/context/spring-context.xsd">
-
-  <!-- common -->
-  <import resource="classpath:common-context.xml" />
-  <!-- db -->
-  <import resource="classpath:db-context.xml" />
-  <!-- business -->
-  <import resource="classpath:business-context.xml" />
-
-  <context:component-scan base-package="jp.co.ha.batch.execute" />
-  <context:component-scan base-package="jp.co.ha.batch.exception" />
-
-  <bean id="messageSource" class="org.springframework.context.support.ResourceBundleMessageSource">
-    <property name="basenames">
-      <list>
-        <value>messages</value>
-        <value>apiErrorCode</value>
-        <value>dashboardErrorCode</value>
-        <value>businessErrorCode</value>
-        <value>commonErrorCode</value>
-        <value>validateErrorCode</value>
-      </list>
-    </property>
-  </bean>
-
-</beans>
