--- ha-batch/src/main/resources/messages.properties
@@ -1,2 +1,0 @@
-batch.result.success=\u6b63\u5e38\u7d42\u4e86
-batch.result.failure=\u7570\u5e38\u7d42\u4e86
