+++ ha-batch/src/test/java/jp/co/ha/batch/BatchApplicationTests.java
@@ -1,0 +1,13 @@
+package jp.co.ha.batch;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class BatchApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
