--- ha-batch/src/test/resources/json/healthInfoRegist_admin.json
@@ -1,19 +1,0 @@
-{
-  "seqUserId": 1,
-  "apiKey": "d9095f2c47b0a651748521077659d03539f4376f17ed79ef070c32d2f6198fb1",
-  "healthInfoRequestDataList" : [
-    {
-      "height": 170.01,
-      "weight": 51.01
-    },
-    {
-      "height": 170.02,
-      "weight": 51.02
-    },
-    {
-      "height": 170.03,
-      "weight": 51.03
-    }
-  ]
-
-}
