+++ ha-pom/pom.xml
@@ -19,7 +19,7 @@
     <org.springframework-version>5.3.7</org.springframework-version>
     <org.spring-data-commons-version>2.5.1</org.spring-data-commons-version>
     <org.spring-security-version>5.5.0</org.spring-security-version>
-    <com.amazonaws-version>1.11.939</com.amazonaws-version>
+    <com.amazonaws-version>1.11.1030</com.amazonaws-version>
     <org.aspectj-version>1.9.6</org.aspectj-version>
     <org.slf4j-version>1.7.30</org.slf4j-version>
     <org.logback-version>1.2.3</org.logback-version>
